pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gh-wikis"
version = "0.1.0"
description = "GitHub Wiki to PDF/EPUB converter"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.3",
    "sqlalchemy>=2.0.27",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.1",
    "pygithub>=2.1.1",
    "minio>=7.2.0",
    "aiohttp>=3.9.3",
    "markdown>=3.6",
    "weasyprint>=61.0",
    "ebooklib>=0.18.0",
    "beautifulsoup4>=4.12.0",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "python-multipart>=0.0.6"
]

# Explicitly define packages for hatchling
[tool.hatch.build.targets.wheel]
packages = ["gh_wikis"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=5.0.0",
    "httpx>=0.26.0",
    "pre-commit>=3.6.0",
    "black>=24.1.1",
    "isort>=5.13.2",
    "mypy>=1.8.0",
]

[tool.pytest]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false